# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  '/api/v1/comments/post/{post_id}':
    get:
      summary: 'comments for one post'
      operationId: commentsForOnePost
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: comments
                  data:
                    comments: []
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: comments
                  data:
                    type: object
                    example:
                      comments: []
      tags:
        - Comment
      security: []
    parameters:
      -
        in: path
        name: post_id
        description: 'The ID of the post.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/v1/files/post/{post_id}':
    get:
      summary: ''
      operationId: getApiV1FilesPostPost_id
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: files
                  data:
                    files:
                      -
                        id: 17
                        post_id: 1
                        path: 'https://source.unsplash.com/random'
                        type: image
                        created_at: '2022-11-16T19:25:31.000000Z'
                        updated_at: '2022-11-16T19:25:31.000000Z'
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: files
                  data:
                    type: object
                    example:
                      files:
                        -
                          id: 17
                          post_id: 1
                          path: 'https://source.unsplash.com/random'
                          type: image
                          created_at: '2022-11-16T19:25:31.000000Z'
                          updated_at: '2022-11-16T19:25:31.000000Z'
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: post_id
        description: 'The ID of the post.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/v1/reacts/post/{post_id}':
    get:
      summary: ''
      operationId: getApiV1ReactsPostPost_id
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: Reacts
                  data:
                    reacts:
                      -
                        id: 1
                        user_id: 40
                        post_id: 1
                        type: 1
                        created_at: '2022-11-16T19:25:32.000000Z'
                        updated_at: '2022-11-16T19:25:32.000000Z'
                      -
                        id: 2
                        user_id: 3
                        post_id: 1
                        type: 3
                        created_at: '2022-11-16T19:25:32.000000Z'
                        updated_at: '2022-11-16T19:25:32.000000Z'
                      -
                        id: 3
                        user_id: 15
                        post_id: 1
                        type: 4
                        created_at: '2022-11-16T19:25:32.000000Z'
                        updated_at: '2022-11-16T19:25:32.000000Z'
                      -
                        id: 4
                        user_id: 13
                        post_id: 1
                        type: 4
                        created_at: '2022-11-16T19:25:32.000000Z'
                        updated_at: '2022-11-16T19:25:32.000000Z'
                      -
                        id: 5
                        user_id: 27
                        post_id: 1
                        type: 4
                        created_at: '2022-11-16T19:25:32.000000Z'
                        updated_at: '2022-11-16T19:25:32.000000Z'
                      -
                        id: 6
                        user_id: 33
                        post_id: 1
                        type: 2
                        created_at: '2022-11-16T19:25:32.000000Z'
                        updated_at: '2022-11-16T19:25:32.000000Z'
                      -
                        id: 7
                        user_id: 32
                        post_id: 1
                        type: 1
                        created_at: '2022-11-16T19:25:32.000000Z'
                        updated_at: '2022-11-16T19:25:32.000000Z'
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Reacts
                  data:
                    type: object
                    example:
                      reacts:
                        -
                          id: 1
                          user_id: 40
                          post_id: 1
                          type: 1
                          created_at: '2022-11-16T19:25:32.000000Z'
                          updated_at: '2022-11-16T19:25:32.000000Z'
                        -
                          id: 2
                          user_id: 3
                          post_id: 1
                          type: 3
                          created_at: '2022-11-16T19:25:32.000000Z'
                          updated_at: '2022-11-16T19:25:32.000000Z'
                        -
                          id: 3
                          user_id: 15
                          post_id: 1
                          type: 4
                          created_at: '2022-11-16T19:25:32.000000Z'
                          updated_at: '2022-11-16T19:25:32.000000Z'
                        -
                          id: 4
                          user_id: 13
                          post_id: 1
                          type: 4
                          created_at: '2022-11-16T19:25:32.000000Z'
                          updated_at: '2022-11-16T19:25:32.000000Z'
                        -
                          id: 5
                          user_id: 27
                          post_id: 1
                          type: 4
                          created_at: '2022-11-16T19:25:32.000000Z'
                          updated_at: '2022-11-16T19:25:32.000000Z'
                        -
                          id: 6
                          user_id: 33
                          post_id: 1
                          type: 2
                          created_at: '2022-11-16T19:25:32.000000Z'
                          updated_at: '2022-11-16T19:25:32.000000Z'
                        -
                          id: 7
                          user_id: 32
                          post_id: 1
                          type: 1
                          created_at: '2022-11-16T19:25:32.000000Z'
                          updated_at: '2022-11-16T19:25:32.000000Z'
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: post_id
        description: 'The ID of the post.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/v1/posts/{post_id}':
    get:
      summary: 'show post'
      operationId: showPost
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: post
                  data:
                    posts:
                      id: 1
                      user_id: 19
                      body: 'Ea quibusdam laboriosam minus expedita quidem iure. Ipsa quisquam impedit incidunt rem fuga. Harum delectus officia quia sed. Omnis voluptatem neque non earum.'
                      created_at: '2022-11-16T19:25:31.000000Z'
                      updated_at: '2022-11-16T19:25:31.000000Z'
                      comments: []
                      files:
                        -
                          id: 17
                          post_id: 1
                          path: 'https://source.unsplash.com/random'
                          type: image
                          created_at: '2022-11-16T19:25:31.000000Z'
                          updated_at: '2022-11-16T19:25:31.000000Z'
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: post
                  data:
                    type: object
                    example:
                      posts:
                        id: 1
                        user_id: 19
                        body: 'Ea quibusdam laboriosam minus expedita quidem iure. Ipsa quisquam impedit incidunt rem fuga. Harum delectus officia quia sed. Omnis voluptatem neque non earum.'
                        created_at: '2022-11-16T19:25:31.000000Z'
                        updated_at: '2022-11-16T19:25:31.000000Z'
                        comments: []
                        files:
                          -
                            id: 17
                            post_id: 1
                            path: 'https://source.unsplash.com/random'
                            type: image
                            created_at: '2022-11-16T19:25:31.000000Z'
                            updated_at: '2022-11-16T19:25:31.000000Z'
      tags:
        - Post
      security: []
    put:
      summary: 'update post'
      operationId: updatePost
      description: ''
      parameters:
        -
          in: query
          name: apiKey
          description: 'Authentication key.'
          example: '{YOUR_AUTH_KEY}'
          required: true
          schema:
            type: string
            description: 'Authentication key.'
            example: '{YOUR_AUTH_KEY}'
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - Post
    delete:
      summary: 'delete post'
      operationId: deletePost
      description: ''
      parameters:
        -
          in: query
          name: apiKey
          description: 'Authentication key.'
          example: '{YOUR_AUTH_KEY}'
          required: true
          schema:
            type: string
            description: 'Authentication key.'
            example: '{YOUR_AUTH_KEY}'
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - Post
    parameters:
      -
        in: path
        name: post_id
        description: 'The ID of the post.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/v1/users/{user_id}/posts':
    get:
      summary: 'show posts for user'
      operationId: showPostsForUser
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: 'Posts for User: Dr. Orlo Bradtke'
                  data:
                    posts:
                      -
                        id: 20
                        user_id: 1
                        body: 'Corporis et est veniam ut iste est. Ducimus atque qui ut quia. Ex rem adipisci ducimus consectetur in laborum.'
                        created_at: '2022-11-16T19:25:31.000000Z'
                        updated_at: '2022-11-16T19:25:31.000000Z'
                      -
                        id: 36
                        user_id: 1
                        body: 'Et similique occaecati illum sunt et non dolores. Qui perferendis autem aspernatur laborum odio. Ut culpa assumenda ea perferendis. Deserunt est non non cupiditate ut.'
                        created_at: '2022-11-16T19:25:31.000000Z'
                        updated_at: '2022-11-16T19:25:31.000000Z'
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Posts for User: Dr. Orlo Bradtke'
                  data:
                    type: object
                    example:
                      posts:
                        -
                          id: 20
                          user_id: 1
                          body: 'Corporis et est veniam ut iste est. Ducimus atque qui ut quia. Ex rem adipisci ducimus consectetur in laborum.'
                          created_at: '2022-11-16T19:25:31.000000Z'
                          updated_at: '2022-11-16T19:25:31.000000Z'
                        -
                          id: 36
                          user_id: 1
                          body: 'Et similique occaecati illum sunt et non dolores. Qui perferendis autem aspernatur laborum odio. Ut culpa assumenda ea perferendis. Deserunt est non non cupiditate ut.'
                          created_at: '2022-11-16T19:25:31.000000Z'
                          updated_at: '2022-11-16T19:25:31.000000Z'
      tags:
        - User
      security: []
    parameters:
      -
        in: path
        name: user_id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: apiKey
      name: apiKey
      in: query
      description: ''
security:
  -
    default: []